Code and project restructure.

+++ CapturePoints.py
@@ -2,7 +2,7 @@ import cv2 as cv
 import numpy as np
 import os
 import glob
-import time
+from lib.ImageOperations import _find_dot
 import json
 
 image_points = [] # format [[camera1_points], [camera1_points], ...] -> timestamp1 = [timestamp1, timestamp2, ...]
@@ -61,7 +61,7 @@ def get_floor_images(preview=False,debug=False):
     global camera_count
     images = []
 
-    image_names = sorted(glob.glob(f'./get_floor_images/test/*.jpg'))
+    image_names = sorted(glob.glob(f'./floor_images/test/*.jpg'))
     image_count = 1
     camera_count = len(image_names)
     if debug:
@@ -69,7 +69,7 @@ def get_floor_images(preview=False,debug=False):
     for fname in image_names:
         print(fname)
         img = cv.imread(fname)
-        img = image_filter(img)
+        # img = image_filter(img)
         if preview:
             cv.imshow(f'{fname}',img)
             key = cv.waitKey(0) & 0xFF
@@ -83,45 +83,6 @@ def get_floor_images(preview=False,debug=False):
     print("Point count:", image_count)
     print("Camera count:", camera_count)
     print("Images shape",images.shape)
-
-def image_filter(image,camera_number=0):
-    '''output: filtered image
-    prerequisites: camera_params distortion'''
-    global camera_params
-    if camera_params is None:
-        print("Camera parameters not found.")
-        quit()
-    image = cv.medianBlur(image, 5)
-    image = cv.threshold(image, 255*0.9, 255, cv.THRESH_BINARY)[1]
-    return image
-
-def _find_dot(img):
-    '''output: image with dot and dot coordinates
-    prerequisites needed: image'''
-    grey = cv.cvtColor(img, cv.COLOR_RGB2GRAY)
-    grey = image_filter(grey)
-    contours,_ = cv.findContours(grey, cv.RETR_TREE, cv.CHAIN_APPROX_SIMPLE)
-    img = cv.drawContours(img, contours, -1, (0,0,255), 4)
-
-    image_points = []
-    for contour in contours:
-        moments = cv.moments(contour)
-        if moments["m00"] != 0:
-            center_x = int(moments["m10"] / moments["m00"])
-            center_y = int(moments["m01"] / moments["m00"])
-            # cv.putText(img, f'({center_x}, {center_y})', (center_x-240,center_y - 15), cv.FONT_HERSHEY_SIMPLEX,2, (0,0,255), 4)
-            # cv.circle(img, (center_x,center_y), 2, (0,255,0), 8)
-            image_points.append([center_x, center_y])
-
-    if len(image_points) > 0:
-        for i in range(len(image_points)):
-            cv.putText(img, str(i), (image_points[i][0]-20,image_points[i][1] - 15), cv.FONT_HERSHEY_SIMPLEX,2, (0,0,255), 4)
-            cv.circle(img, (image_points[i][0],image_points[i][1]), 2, (0,255,0), 8)
-
-    if len(image_points) == 0:
-        image_points = [[None, None]]
-
-    return img, image_points
 
 def capture_pose_points(preview=False,debug=False):
     '''output: saves points to points_json
@@ -149,47 +110,53 @@ def capture_pose_points(preview=False,debug=False):
     for j in range(0, image_count):
         processed_images = []
         calculated_points = np.zeros((camera_count, 2))
+        skip = False
         
         for i in range(0, camera_count):
             image, detected_points = _find_dot(images[i][j])
             
-            # Display image with detected points
-            display_image = image.copy()
+            if not detected_points == [[None, None]]:
+                # Display image with detected points
+                display_image = image.copy()
 
-            # Clear prev clicked point
-            clicked_point = None
+                # Clear prev clicked point
+                clicked_point = None
 
-            
-            while clicked_point is None:
-                cv.imshow('Preview', display_image)
-                cv.putText(display_image, "Click on the desired point", (10, 30), 
-                           cv.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
                 
-                key = cv.waitKey(10)
-                if key == 27:  # ESC key to exit
-                    cv.destroyAllWindows()
-                    quit()
-            
-            # Find the closest detected point to where the user clicked
-            min_dist = float('inf')
-            closest_point = None
-            idx = -1
-            
-            for k in range(len(detected_points)):
-                dist = ((detected_points[k][0] - clicked_point[0]) ** 2 + (detected_points[k][1] - clicked_point[1]) ** 2) ** 0.5
-                if dist < min_dist:
-                    min_dist = dist
-                    closest_point = detected_points[k]
-                    idx = k
-            
-            if closest_point is None:
-                print("No points detected near click. Please try again.")
-                i -= 1  # Retry this camera
-                continue
+                while clicked_point is None:
+                    cv.imshow('Preview', display_image)
+                    cv.putText(display_image, "Click on the desired point", (10, 30), 
+                            cv.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
+                    
+                    key = cv.waitKey(10)
+                    if key == 27:  # ESC key to exit
+                        cv.destroyAllWindows()
+                        quit()
+                # Find the closest detected point to where the user clicked
+                min_dist = float('inf')
+                closest_point = None
+                idx = -1
+
+                for k in range(len(detected_points)):
+                    dist = ((detected_points[k][0] - clicked_point[0]) ** 2 + (detected_points[k][1] - clicked_point[1]) ** 2) ** 0.5
+                    if dist < min_dist:
+                        min_dist = dist
+                        closest_point = detected_points[k]
+                        idx = k
                 
-            # Use the closest detected point
+                if closest_point is None:
+                    print("No points detected near click. Please try again.")
+                    i -= 1  # Retry this camera
+                    continue
+                    
+                # Use the closest detected point
+                image_point = np.array(closest_point)
+            else:
+                print("No points detected.Skipping the set.")
+                image_point = [None, None]
+                skip = True
+                break
             print(f"{idx} is clicked!")
-            image_point = np.array(closest_point)
             calculated_points[i] = image_point
             processed_images.append(image)
             if len(image_point) != 2:
@@ -197,25 +164,26 @@ def capture_pose_points(preview=False,debug=False):
                 quit()
             else:
                 image_point = image_point[0]
-        if preview:
-            image = cv.resize(image, (int(image.shape[1] * 0.25), int(image.shape[0] * 0.25)))
-            image = np.hstack([processed_images[0], processed_images[1]])
-            cv.imshow("Preview", image)
-            key = cv.waitKey(0) & 0xFF
-            if key == ord(' '):
-                print("Saving points...")
+        if not skip:
+            if preview:
+                image = cv.resize(image, (int(image.shape[1] * 0.25), int(image.shape[0] * 0.25)))
+                image = np.hstack([processed_images[0], processed_images[1]])
+                cv.imshow("Preview", image)
+                key = cv.waitKey(0) & 0xFF
+                if key == ord(' '):
+                    print("Saving points...")
+                    image_points.append(calculated_points.tolist())
+                if key == ord('x'):
+                    print("Skipping points...")
+                if key == ord('q'):
+                    print("Exiting...")
+                    cv.destroyAllWindows()
+                    quit()
+            else:
                 image_points.append(calculated_points.tolist())
-            if key == ord('x'):
-                print("Skipping points...")
-            if key == ord('q'):
-                print("Exiting...")
-                cv.destroyAllWindows()
-                quit()
-        else:
-            image_points.append(calculated_points.tolist())
-        
-        if debug:
-            print("Image points for capture", i, ":", image_points)
+            
+            if debug:
+                print("Image points for capture", i, ":", image_points)
     with open(points_json, "w") as file:
         json.dump(image_points, file)
